# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Ana Portfolio API
  description: >-
    API do portfólio: conteúdo público, editor autenticado, aparência e
    assistente.


    ### Como usar (passo a passo)

    1. Teste a API em **GET /api/v1/health**.

    2. Leia dados com os **GET** 🔓 (públicos) e anote o campo `revision`.

    3. Para alterar dados, faça login em **POST /api/v1/auth/login** e copie o
    `accessToken`.

    4. Clique em **Authorize** (topo da página), cole o token e confirme.

    5. Use os endpoints 🔒 (POST, PUT, DELETE) enviando `expectedRevision`: a
    **versão atual dos dados**, igual ao `revision` do último GET.


    ### Erros comuns

    - **400**: campo faltando ou em formato errado.

    - **401**: sem login ou token vencido.

    - **404**: o `id` não existe.

    - **409**: `expectedRevision` desatualizado. Faça um novo GET e tente de
    novo.
  version: 1.0.0
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/v1/auth/login:
    post:
      summary: Faz login e recebe o token
      tags:
        - Auth
      description: >-
        🔓 **Público** — não precisa de login.


        Envie usuário e senha do editor. Copie o `accessToken` retornado e cole
        no botão **Authorize** (topo da página) para liberar os endpoints 🔒.
        Cole só o token, sem a palavra "Bearer".
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              additionalProperties: false
              properties:
                username:
                  type: string
                  minLength: 1
                  description: Usuário do editor (definido em ADMIN_USERNAME).
                  example: admin
                password:
                  type: string
                  minLength: 1
                  description: Senha do editor (definida em ADMIN_PASSWORD).
                  example: sua-senha
            example:
              username: admin
              password: sua-senha
      responses:
        "200":
          description: Login feito. Use o `accessToken` no botão Authorize.
          content:
            application/json:
              schema:
                description: Login feito. Use o `accessToken` no botão Authorize.
                type: object
                required:
                  - accessToken
                  - tokenType
                  - expiresIn
                properties:
                  accessToken:
                    type: string
                    description: Token JWT. Cole no botão Authorize.
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.exemplo.assinatura
                  tokenType:
                    type: string
                    description: Tipo do token (sempre "Bearer").
                    enum:
                      - Bearer
                    example: Bearer
                  expiresIn:
                    type: string
                    description: Tempo de validade do token.
                    example: 15m
        "400":
          description: "Dados inválidos: algum campo está faltando ou fora do formato."
          content:
            application/json:
              schema:
                description: "Dados inválidos: algum campo está faltando ou fora do formato."
                type: object
                required:
                  - code
                  - message
                properties:
                  code:
                    type: string
                    description: Código do erro.
                    example: VALIDATION_ERROR
                  message:
                    type: string
                    description: Explicação do erro.
                    example: A requisição não atende ao contrato da API.
                  details:
                    type: object
                    additionalProperties: true
                    description: Informações extras (opcional).
        "401":
          description: Usuário ou senha incorretos.
          content:
            application/json:
              schema:
                description: Usuário ou senha incorretos.
                type: object
                required:
                  - code
                  - message
                properties:
                  code:
                    type: string
                    example: INVALID_CREDENTIALS
                  message:
                    type: string
                    example: Usuário ou senha inválidos.
        "503":
          description: Login ainda não configurado no servidor.
          content:
            application/json:
              schema:
                description: Login ainda não configurado no servidor.
                type: object
                required:
                  - code
                  - message
                properties:
                  code:
                    type: string
                    example: AUTH_NOT_CONFIGURED
                  message:
                    type: string
                    example: Configure ADMIN_USERNAME, ADMIN_PASSWORD e JWT_SECRET antes de usar a
                      autenticação.
  /api/v1/auth/me:
    get:
      summary: Mostra quem está logado
      tags:
        - Auth
      description: >-
        🔒 **Requer Bearer Token** — faça login em `POST /api/v1/auth/login` e
        cole o `accessToken` no botão **Authorize**.


        Serve para testar se o token está valendo: retorna o usuário dono do
        token.
      security: &a1
        - bearerAuth: []
      responses:
        "200":
          description: Token válido.
          content:
            application/json:
              schema:
                description: Token válido.
                type: object
                required:
                  - username
                properties:
                  username:
                    type: string
                    description: Usuário logado.
                    example: admin
        "401":
          description: "Sem login: envie o Bearer Token (botão Authorize)."
          content:
            application/json:
              schema:
                description: "Sem login: envie o Bearer Token (botão Authorize)."
                type: object
                required:
                  - code
                  - message
                properties:
                  code:
                    type: string
                    description: Código do erro.
                    example: UNAUTHORIZED
                  message:
                    type: string
                    description: Explicação do erro.
                    example: Autenticação obrigatória ou token inválido.
                  details:
                    type: object
                    additionalProperties: true
                    description: Informações extras (opcional).
  /api/v1/auth/logout:
    post:
      summary: Encerra a sessão
      tags:
        - Auth
      description: >-
        🔒 **Requer Bearer Token** — faça login em `POST /api/v1/auth/login` e
        cole o `accessToken` no botão **Authorize**.


        Não tem corpo. O token não é invalidado no servidor: depois de chamar,
        descarte o token (no Swagger, use **Logout** no botão Authorize).
      security: *a1
      responses:
        "204":
          description: Sessão encerrada (resposta vazia).
        "401":
          description: "Sem login: envie o Bearer Token (botão Authorize)."
          content:
            application/json:
              schema:
                description: "Sem login: envie o Bearer Token (botão Authorize)."
                type: object
                required:
                  - code
                  - message
                properties:
                  code:
                    type: string
                    description: Código do erro.
                    example: UNAUTHORIZED
                  message:
                    type: string
                    description: Explicação do erro.
                    example: Autenticação obrigatória ou token inválido.
                  details:
                    type: object
                    additionalProperties: true
                    description: Informações extras (opcional).
...
servers:
  - url: http://localhost:3333
    description: Desenvolvimento local
...